Add files via upload
In this project, we are carrying out Zero shot recognition using varying forms of feature extraction and training models, the goal being to classify images of different animals, using descriptors. For my submission, I have chosen convolutional neural networks (CNNs). 

I made the image data accessible to our imds; I did this by means of a for-loop that iterates through the images, storing them into two matrices called trainingPaths and testingPaths, which correspond to the TrainingImages and testingImages files susch that they remain subject to our train test split of 80/20. This is to say that 80 percent of our 50 classes are used to train the model and the remaining are the testing classes and hence fed into separate matrices. This will result in a 40*85 training matrix  and 10*85 testing matrix.